test: fix ut
