+++ test/ut/spec/series/custom.test.ts
@@ -20,7 +20,7 @@
 import { EChartsType } from '../../../../src/echarts';
 import { createChart } from '../../core/utHelper';
 import { ZRColor } from '../../../../src/util/types';
-import { CustomSeriesRenderItemAPI, CustomSeriesRenderItemParams } from '../../../../src/chart/custom';
+import { CustomSeriesRenderItemAPI, CustomSeriesRenderItemParams } from '../../../../src/chart/custom/install';
 
 
 describe('custom_series', function () {
